+++ app/models/checklist/pdf/index_content.rb
@@ -8,39 +8,8 @@ module Checklist::Pdf::IndexContent
     kingdom(tex, fetcher, 'FAUNA')
     fetcher = Checklist::Pdf::IndexFetcher.new(@plantae_query)
     kingdom(tex, fetcher, 'FLORA')
-    tex << annotations_key
-  end
-
-  def annotations_key
-    tex = "\\parindent 0in"
-    tex << "\\cpart{Annotations key}\n"
-    cites = Designation.find_by_name(Designation::CITES)
-    taxa_with_index_annotations = MTaxonConcept.
-      select("taxon_concepts_mview.*, annotations.symbol, annotations.full_note_en").
-      joins( <<-SQL
-        JOIN listing_changes
-        ON listing_changes.taxon_concept_id = taxon_concepts_mview.id
-        AND listing_changes.is_current
-        JOIN change_types
-        ON change_types.id = listing_changes.change_type_id
-        AND designation_id = #{cites.id}
-        AND change_types.name = 'ADDITION'
-        JOIN annotations
-        ON listing_changes.annotation_id = annotations.id
-        AND annotations.display_in_index
-        AND annotations.symbol IS NOT NULL
-      SQL
-      ).
-      order("annotations.symbol::INT")
-
-    tex << "\\section*{Annotations not preceded by \"\\#\"}\n"
-    taxa_with_index_annotations.each do |tc|
-        box_colour = (tc.kingdom_name == 'Animalia' ? 'orange' : 'green')
-        tex << "\\cfbox{#{box_colour}}{\\superscript{#{tc['symbol']}} \\textbf{#{taxon_name_at_rank(tc)}}}\n\n"
-        tex << "#{LatexToPdf.html2latex(tc['full_note_en'])}\n\n"
-      end
-    tex << "\\parindent -0.1in"
-    tex
+    ak = Checklist::Pdf::IndexAnnotationsKey.new
+    tex << ak.annotations_key
   end
 
   def kingdom(tex, fetcher, kingdom_name)
@@ -83,20 +52,6 @@ module Checklist::Pdf::IndexContent
       res += " (#{tc.class_name})" unless tc.class_name.blank?
     end
     res += common_names_with_lng_initials(tc)
-    res
-  end
-
-  def taxon_name_at_rank(taxon_concept)
-    res = if ['FAMILY','SUBFAMILY','ORDER','CLASS'].include? taxon_concept.rank_name
-      LatexToPdf.escape_latex(taxon_concept.full_name.upcase)
-    else
-      if ['SPECIES', 'SUBSPECIES', 'GENUS'].include? taxon_concept.rank_name
-        "\\textit{#{LatexToPdf.escape_latex(taxon_concept.full_name)}}"
-      else
-        LatexToPdf.escape_latex(taxon_concept.full_name)
-      end
-    end
-    res += " #{LatexToPdf.escape_latex(taxon_concept.spp)}" if taxon_concept.spp
     res
   end