+++ app/serializers/checklist/autocomplete_taxon_concept_serializer.rb
@@ -1,0 +1,7 @@
+class Checklist::AutocompleteTaxonConceptSerializer < ActiveModel::Serializer
+  attributes :id, :full_name, :rank_name, :matching_names
+
+  def rank_name
+    object.rank_display_name
+  end
+end