+++ lib/modules/downloads_cache.rb
@@ -35,6 +35,7 @@ module DownloadsCache
   # for admin purposes
   def self.clear
     clear_dirs(DOWNLOAD_DIRS)
+    Download.delete_all
   end
 
   def self.clear_cites_listings
@@ -128,27 +129,28 @@ module DownloadsCache
       Checklist::Csv,
       Checklist::Json
     ]
+    ['en', 'es', 'fr'].each do |locale|
+      # full download parameters
+      params = {
+        show_synonyms: "1",
+        show_author: "1",
+        show_english: "1",
+        show_spanish: "1",
+        show_french: "1",
+        intro: "1",
+        locale: locale
+      }
 
-    # full download parameters
-    params = {
-      show_synonyms: "1",
-      show_author: "1",
-      show_english: "1",
-      show_spanish: "1",
-      show_french: "1",
-      locale: "en",
-      format: "json"
-    }
-
-    modules.each do |m|
-      elapsed_time = Benchmark.realtime do
-        puts m::Index.new(params).generate
+      modules.each do |m|
+        elapsed_time = Benchmark.realtime do
+          puts m::Index.new(params).generate
+        end
+        puts "#{Time.now} #{m}::Index download #{locale} generated in #{elapsed_time}s"
+        elapsed_time = Benchmark.realtime do
+          puts m::History.new(params).generate
+        end
+        puts "#{Time.now} #{m}::History download #{locale} generated in #{elapsed_time}s"
       end
-      puts "#{Time.now} #{m}::Index download generated in #{elapsed_time}s"
-      elapsed_time = Benchmark.realtime do
-        puts m::History.new(params).generate
-      end
-      puts "#{Time.now} #{m}::History download generated in #{elapsed_time}s"
     end
   end