+++ spec/models/checklist/timeline_spec.rb
@@ -81,8 +81,8 @@ describe Checklist::Timeline do
     let(:subject){ ttc.timelines.last }
 
     specify{ subject.timeline_intervals.count.should == 3 }
-    pending{ subject.timeline_intervals.last.end_pos.should < 1 }
-    pending{ subject.timeline_events.count.should == 4 }
+    specify{ subject.timeline_intervals.last.end_pos.should < 1 }
+    specify{ subject.timeline_events.count.should == 4 }
   end
 
   context "when deleted and then readded" do
@@ -109,8 +109,100 @@ describe Checklist::Timeline do
     let(:ttc){ Checklist::TimelinesForTaxonConcept.new(tc)}
     let(:subject){ ttc.timelines.first }
 
-    pending{ subject.timeline_intervals.count.should == 2 }
-    pending{ subject.timeline_events.count.should == 3 }
+    specify{ subject.timeline_intervals.count.should == 2 }
+    specify{ subject.timeline_events.count.should == 3 }
+    specify{ subject.timeline_intervals[0].end_pos.should == subject.timeline_intervals[1].start_pos }
+  end
+
+  context "when reservation withdrawn" do
+    let(:tc){
+      tc = create_cites_eu_species
+      create_cites_I_addition(
+        :taxon_concept => tc,
+        :effective_at => '1975-06-06',
+        :is_current => true
+      )
+      cnt = create(:geo_entity, geo_entity_type: country_geo_entity_type)
+      r1 = create_cites_I_reservation(
+        :taxon_concept => tc,
+        :effective_at => '1975-06-06',
+        :is_current => false
+      )
+      create(
+        :listing_distribution,
+        :geo_entity => cnt,
+        :listing_change => r1,
+        :is_party => true
+      )
+      w = create_cites_I_reservation_withdrawal(
+        :taxon_concept => tc,
+        :effective_at => '1976-06-07',
+        :is_current => false
+      )
+      create(
+        :listing_distribution,
+        :geo_entity => cnt,
+        :listing_change => w,
+        :is_party => true
+      )
+      Sapi::StoredProcedures.rebuild_cites_taxonomy_and_listings
+      MTaxonConcept.find(tc.id)
+    }
+    let(:ttc){ Checklist::TimelinesForTaxonConcept.new(tc)}
+    let(:subject){ ttc.timelines.first.timelines.first }
+
+    specify{ puts subject.timeline_events.inspect; puts subject.timeline_intervals.inspect; subject.timeline_intervals.count.should == 1 }
+    specify{ subject.timeline_events.count.should == 2 }
+    specify{ subject.timeline_intervals[0].end_pos.should == subject.timeline_events[1].pos }
+  end
+
+  context "when reservation withdrawn and then readded" do
+    let(:tc){
+      tc = create_cites_eu_species
+      cnt = create(:geo_entity, geo_entity_type: country_geo_entity_type)
+      r1 = create_cites_III_reservation(
+        :taxon_concept => tc,
+        :effective_at => '1975-06-06',
+        :is_current => false
+      )
+      create(
+        :listing_distribution,
+        :geo_entity => cnt,
+        :listing_change => r1,
+        :is_party => true
+      )
+      w = create_cites_III_reservation_withdrawal(
+        :taxon_concept => tc,
+        :effective_at => '1976-06-07',
+        :is_current => false
+      )
+      create(
+        :listing_distribution,
+        :geo_entity => cnt,
+        :listing_change => w,
+        :is_party => true
+      )
+      r2 = create_cites_III_reservation(
+        :taxon_concept => tc,
+        :effective_at => '1977-06-06',
+        :is_current => true
+      )
+      create(
+        :listing_distribution,
+        :geo_entity => cnt,
+        :listing_change => r2,
+        :is_party => true
+      )
+      Sapi::StoredProcedures.rebuild_cites_taxonomy_and_listings
+      MTaxonConcept.find(tc.id)
+    }
+    let(:ttc){ Checklist::TimelinesForTaxonConcept.new(tc)}
+    let(:subject){ ttc.timelines.last.timelines.first }
+
+    specify{ subject.timeline_intervals.count.should == 2 }
+    specify{ subject.timeline_events.count.should == 3 }
+    specify{ subject.timeline_intervals[0].end_pos.should == subject.timeline_events[1].pos }
+    specify{ subject.timeline_intervals[1].end_pos.should == 1 }
   end
 
   context "when added multiple times" do
@@ -136,6 +228,8 @@ describe Checklist::Timeline do
       subject.timeline_events.map(&:change_type_name).should ==
         ['ADDITION', 'AMENDMENT']
     }
+    specify{ subject.timeline_intervals.count.should == 2 }
+    specify{ subject.timeline_intervals[1].end_pos.should == 1 }
   end
 
 end